Move RPacketTracker class to a new file packet_trackers.py
